# Fix semaphore spinning fix to not spin when count available