A recent fix around SemaphoreSlim's spinning has led to it spinning even when spinning isn't necessary, resulting in measurable overheads when trying to acquire a semaphore that has count available.  This fixes that by putting back the check to avoid spinning if count is known to be available.

Fixes https://github.com/dotnet/coreclr/issues/14980
cc: @kouvel, @benaadams 

Timings for doing wait/release repeatedly on a semaphore with count available...

Before:

|    Method |       Mean |     Error |    StdDev |
---------- |-----------:|----------:|----------:|
| WaitAsync() |   413.5 ns |  6.752 ns |  6.316 ns |
|  Wait(0) | 1,050.8 ns | 21.816 ns | 25.124 ns |
|      Wait() | 1,090.8 ns | 21.090 ns | 24.288 ns |

After:

|    Method |     Mean |    Error |    StdDev |
---------- |---------:|---------:|----------:|
| WaitAsync() | 421.6 ns | 8.478 ns | 17.883 ns |
|  Wait(0) | 465.4 ns | 8.847 ns | 10.189 ns |
|      Wait() | 457.3 ns | 6.808 ns |  6.368 ns |